Refactor connection initialization and export Conn struct
The previous code didn't really work, and it lacked tests. This commit
makes several changes to the previous version and adds some necessary
tests (which are far from perfect). The first important change is that
the internal `conn` struct now is exported as `Conn`, which will allow
users to cast to `*viaproxy.Conn` and access the additional
ProxyAddr() method that returns the proxy address.

The proxied connection initialization code was also rewrote almost
from scratch, and while I'm still not very happy about it (I'm
thinking how we can get rid of the buffered connection) it's working
better than before, IMO.