Return concrete type instead of an interface
This is more idiomatic in Go. Also it saves the caller the need to
typecast the returned object of calling viaproxy.Wrap to a
viaproxy.Conn and thus giving automatically access to methods like
ProxyAddr. And given that viaproxy.Conn implements the net.Conn
interface, it can be passed as an argument to any other methods that
accept a net.Conn.